chore: convert PlayerScore from int to float to enable divisions in the future

+++ Source/AimTrainer/Gameplay/AimTrainerGameMode.h
@@ -58,7 +58,7 @@ public:
 private:
 	FTimerHandle SessionTimerHandle;
 
-	int32 PlayerScore = 0;
+	float PlayerScore = 0.f;
 
 	UPROPERTY()
 	UHUDWidget* HUDWidget;
